Refactor providers into separate files
